chore(front): Move form wrapper to form element directory
Rename directory to form-element
