+++ client/src/components/apply-form/index.ts
@@ -2,4 +2,3 @@ export * from "./area-control";
 export * from "./input-control";
 export * from "./radio-control";
 export * from "./title";
-export * from "./wrapper";
