+++ client/src/components/form-element/type/model.ts
@@ -1,0 +1,34 @@
+import type { ComponentProps, ReactNode } from "react";
+import type {
+  FieldError,
+  DeepPartial,
+  FieldValues,
+  SubmitHandler,
+} from "react-hook-form";
+import type { z } from "zod";
+
+export type TextInputProperties = ComponentProps<"input"> & {
+  inputStyles?: string;
+  error?: FieldError;
+};
+
+export type AreaInputProperties = ComponentProps<"textarea"> & {
+  inputStyles?: string;
+  error?: FieldError;
+};
+
+export type WrapperProperties<T extends FieldValues = never> = {
+  defaultValues: (DeepPartial<T> | undefined) & T;
+  children: ReactNode;
+  onSubmit: SubmitHandler<T>;
+  schema: z.Schema;
+  wrapperStyle?: string;
+};
+
+export type RadioInputProperties = Omit<
+  ComponentProps<"input">,
+  "id" | "type"
+> & {
+  id: string;
+  value: string;
+};
