+++ client/src/pages/apply-form/type/model.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { schema } from "~/components/apply-form/type/schema";
+import { schema } from "~/pages/apply-form/type/schema";
 
 export type ApplyFormPageProperties = {
   eventId: string;
